requirements.txt
Django>=4.2.6
djangorestframework>=3.14.0
django-filter>=23.3
pytest>=7.4.3